refactor: separate canvas block item
