+++ packages/ui/src/components/cms/page-builder/CanvasItem.d.ts
@@ -17,5 +17,6 @@ type Props = {
     device?: DevicePreset;
 };
 declare const CanvasItem: import("react").NamedExoticComponent<Props>;
+export type { Props as CanvasItemProps };
 export default CanvasItem;
 //# sourceMappingURL=CanvasItem.d.ts.map
